Убрал компоненты на классах, вернул функциональные. Заработала передача данных из дочек в родительские через родительский state
