+++ src/components/App.js
@@ -13,40 +13,40 @@ import { useEffect, useState } from 'react';
 
 
 function App({greeting}) {  
-const [messageList, setMessageList] = useState([]);
-const [inputMessage, setInputMessage] = useState('');
+/* const [messageList, setMessageList] = useState([]);
+const [inputMessage, setInputMessage] = useState('');*/
 const adminMessage = {
   author: 'Admin',
   text: 'Your message is accepted!'
 };
 
 
-const sendMessage = (event) => {
-  event.preventDefault();
-  setMessageList([...messageList, inputMessage]);
-  setInputMessage({
-    author: '',
-    text: '',});
-}
+// const sendMessage = (event) => {
+//   event.preventDefault();
+//   setMessageList([...messageList, inputMessage]);
+//   setInputMessage({
+//     author: '',
+//     text: '',});
+// }
 
-const handleChange = (event) => {
-  setInputMessage({
-    author: 'User',
-    text: event.target.value,
-  }); 
-}
+// const handleChange = (event) => {
+//   setInputMessage({
+//     author: 'User',
+//     text: event.target.value,
+//   }); 
+// }
 
-useEffect(() => {
-    setTimeout(() => {
-      console.log('setTimeout: ', messageList[messageList.length - 1]);
-      if ( messageList.length !== 0 && messageList[messageList.length - 1].author !== 'Admin') {
-        setMessageList([...messageList, adminMessage]);
-        setInputMessage({
-          author: '',
-          text: '',});
-        } 
-    }, 1000)
-}, [messageList])
+// useEffect(() => {
+//     setTimeout(() => {
+//       console.log('setTimeout: ', messageList[messageList.length - 1]);
+//       if ( messageList.length !== 0 && messageList[messageList.length - 1].author !== 'Admin') {
+//         setMessageList([...messageList, adminMessage]);
+//         setInputMessage({
+//           author: '',
+//           text: '',});
+//         } 
+//     }, 1000)
+// }, [messageList])
 
   return ( 
     <div className="App" >
