+++ src/components/MessageBox.js
@@ -1,31 +1,54 @@
 import * as React from 'react';
+import { useEffect, useState } from 'react';
 import Box from '@mui/material/Box';
 import Message from './Message';
 import InputMessage from './InputMessage';
 
-//TODO дописать класс
-export default class MessageBox extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {messageList: [], inputMessage: ''};
-    console.log('props: ',this.state);
-    this.sendMessage = this.sendMessage.bind(this);
-  }
 
-  sendMessage(event) {
-    // event.preventDefault();
-    console.log('this.state.messageList: ', this.state);
-    this.setState({messageList: [...this.state.messageList, this.state.inputMessage]});
-    // this.setState({inputMessage: {author: '', text: ''}});
-  }
+
+
+export default function MessageBox() {
+  const [messageList, setMessageList] = useState([]);
   
-  handleChange(event) {
-    this.setState({inputMessage: {author: 'User', text: event.target.value}}); 
-  }
 
-  componentDidUpdate(props) {
+  return (
+    <Box
+      sx={{
+        width: '80%',
+        backgroundColor: 'primary.ligth'
+      }}
+    >
+      <h2>MessageBox</h2>
+        <InputMessage messageList={messageList} setMessageList={setMessageList} />
+        <Message list={messageList}/>
+    </Box>
 
-  }
+  )
+} 
+
+//TODO дописать класс
+// export default class MessageBox extends React.Component {
+//   constructor(props) {
+//     super(props);
+//     this.state = {messageList: [], inputMessage: ''};
+//     console.log('props: ',this.state);
+//     this.sendMessage = this.sendMessage.bind(this);
+//   }
+
+//   sendMessage(event) {
+//     // event.preventDefault();
+//     console.log('this.state.messageList: ', this.state);
+//     this.setState({messageList: [...this.state.messageList, this.state.inputMessage]});
+//     // this.setState({inputMessage: {author: '', text: ''}});
+//   }
+  
+//   handleChange(event) {
+//     this.setState({inputMessage: {author: 'User', text: event.target.value}}); 
+//   }
+
+//   componentDidUpdate(props) {
+
+//   }
 //   componentDidUpdate(() => {
 //     setTimeout(() => {
 //       if ( this.messageList.length !== 0 && this.messageList[this.messageList.length - 1].author !== 'Admin') {
@@ -37,24 +60,24 @@ export default class MessageBox extends React.Component {
 //     }, 1000)
 // },[this.messageList])
 
-render() {
-  return (
-    <Box
-      sx={{
-        width: '80%',
-        backgroundColor: 'primary.ligth'
-      }}
-    >
-        {/* <InputMessage send={this.sendMessage} onChange={this.handleChange}/> */}
-        {/* <Message list={this.state.messageList}/> */}
-  </Box>
-  )
+// render() {
+//   return (
+//     <Box
+//       sx={{
+//         width: '80%',
+//         backgroundColor: 'primary.ligth'
+//       }}
+//     >
+//         {/* <InputMessage send={this.sendMessage} onChange={this.handleChange}/> */}
+//         {/* <Message list={this.state.messageList}/> */}
+//   </Box>
+//   )
   
-}
+// }
 
 
 
-}
+// }
 
 
 // export default function MessageBox() {
